# api/api.yml
...
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: GPL v3
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/serviceaccounts:
    post:
      summary: create a new service account
      operationId: createServiceAccount
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccount"
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created service account can be found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: get service accounts for a project
      operationId: getServiceAccountsForProject
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: the supply chain
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceAccount"
        '404':
          description: not found
        '500':
          $ref: '#/components/responses/Error'
  /serviceaccounts:
    get:
      summary: get service accounts
      operationId: getServiceAccounts
      tags:
        - service account
      responses:
        '200':
          description: the supply chain
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceAccount"
        '404':
          description: not found
        '500':
          $ref: '#/components/responses/Error'
  /serviceaccounts/me/activekey:
    get:
      summary: return the active key pair
      operationId: getServiceAccountKey
      tags:
        - service account
      responses:
        '200':
          description: active key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyPair"
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: no active key exists
        '500':
          $ref: '#/components/responses/Error'
  /projects/{projectId}/serviceaccounts/{serviceAccountId}:
    get:
      summary: find service account by id
      operationId: getServiceAccountById
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: serviceAccountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: service account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: the service account is not found
        '500':
          $ref: '#/components/responses/Error'
    put:
      summary: update service account by id
      operationId: updateServiceAccountById
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: serviceAccountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccount"
      responses:
        '200':
          description: updated service account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceAccount"
        '404':
          description: the service account is not found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    delete:
      summary: deletes a service account
      operationId: deleteServiceAccount
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: serviceAccountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '204':
          description: deleted service account
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: the service account is not found
        '500':
          $ref: '#/components/responses/Error'
  /projects/{projectId}/serviceaccounts/{serviceAccountId}/key:
    post:
      summary: create a new active key the old key will be made inactive
      operationId: createServiceAccountKeyById
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: serviceAccountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAccountKeyPair"
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyPair"
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created key can be found
        '404':
          description: the service account is not found
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: return the active key pair
      operationId: getServiceAccountKeyById
      tags:
        - service account
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: serviceAccountId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyPair"
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: no active key exists
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    SHA256Hash:
      type: string
      minLength: 64
      maxLength: 64
      pattern: '^[0-9a-f]*$'
    UUID:
      type: string
      format: uuid
...
    NameString:
      type: string
      minLength: 1
      maxLength: 63
      pattern: '^([a-z]|[a-z][a-z0-9-]*[a-z0-9])?$'
    ByteString:
      type: string
      maxLength: 4096
      format: byte
...
    KeyPair:
      type: object
      required:
        - encryptedPrivateKey
        - keyId
        - publicKey
      properties:
        encryptedPrivateKey:
          $ref: '#/components/schemas/ByteString'
        keyId:
          $ref: '#/components/schemas/KeyId'
        publicKey:
          $ref: '#/components/schemas/ByteString'
    ServiceAccountKeyPair:
      type: object
      required:
        - encryptedPrivateKey
        - passphrase
        - keyId
        - publicKey
      properties:
        encryptedPrivateKey:
          $ref: '#/components/schemas/ByteString'
        passphrase:
          type: string
          maxLength: 80
          pattern: '^(?=.*?[a-zA-Z\d])[\w@#&*+=$-]+$'
        keyId:
          $ref: '#/components/schemas/KeyId'
        publicKey:
          $ref: '#/components/schemas/ByteString'
          #writeOnly: true
          #description: used for authentication of a service account via http basic authentication
...
    KeyId:
      $ref: "#/components/schemas/SHA256Hash"
...
    ServiceAccount:
      type: object
      required:
        - name
        - projectId
      properties:
        id:
          $ref: "#/components/schemas/UUID"
          readOnly: true
        providerSubject:
          type: string
          maxLength: 1024
          readOnly: true
        name:
          $ref: "#/components/schemas/NameString"
        projectId:
          $ref: "#/components/schemas/UUID"
        activeKeyPair:
          $ref: "#/components/schemas/KeyPair"
        inactiveKeyPairs:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/KeyPair"
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string
...